修复bug:https://github.com/zhongjhATC/Retrofit2Rxjava2Download/issues/1

+++ retrofit2rxjava2downloadlib/src/main/java/com/zhongjh/retrofitdownloadlib/http/DownloadHandler.java
@@ -67,7 +67,7 @@ class DownloadHandler {
         Message message = mHandler.obtainMessage();
         message.what = WHAT_UPDATE;
         Bundle data = new Bundle();
-        data.putFloat(PROGRESS,  progress);
+        data.putInt(PROGRESS,  progress);
         message.setData(data);
         mHandler.sendMessage(message);
     }
